pyproject.toml
[project]
name = "simstack-model"
version = "0.1.1"
description = "Examples and initial runner tests for Simstack II"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Wolfgang Wenzel", email = "wolfgang.wenzel@kit.edu" }
]
dependencies = [
    "pip==25.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "simstack @ git+https://github.com/simstack/simstack.git@main",
    "scikit-learn",
    "pandas-stubs~=3.0.0",
]

[project.entry-points."simstack.modules"]
example_modules = "examples"
application_modules = "applications"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["examples", "applications"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
